fix bad type limit

+++ src/live_trading_indicators/datasources/binance.py
@@ -108,7 +108,7 @@ def online_request(api_url, symbol, timeframe, start_time, end_time):
     start_time_int = start_time.astype(np.int64)
     end_time_int = end_time.astype(np.int64)
 
-    limit = (end_time - start_time) // timeframe.value + 1
+    limit = int((end_time - start_time).astype(np.int64)) // timeframe.value + 1
     request_url = f'{api_url}klines?symbol={symbol.upper()}&interval={timeframe!s}&startTime={start_time_int}&endTime={end_time_int}&limit={limit}'
 
     logging.debug(f'bars request {request_url}')
